Refactor component directory structure
